Add files via upload
Allows you to place formulas as note tags to adjust damage output of elements. For example you can add exactly 100 points of damage to the final total of a skill. you can use this on armor and weapons as well. it also has a note tag for reversing elemental damage (elemental absorption).